# swagger.yaml
components:
  schemas:
    CreateItem-Input:
      properties:
        item_name:
          title: Item Name
          type: string
      required:
      - item_name
      title: CreateItem
      type: object
    CreateItem-Output:
      properties:
        item_id:
          title: Item Id
          type: integer
        item_name:
          title: Item Name
          type: string
      required:
      - item_id
      - item_name
      title: CreateItem
      type: object
    DeleteItem:
      properties:
        item_id:
          title: Item Id
          type: integer
        item_name:
          title: Item Name
          type: string
      required:
      - item_id
      - item_name
      title: DeleteItem
      type: object
    GetItem:
      properties:
        item_id:
          title: Item Id
          type: integer
        item_name:
          title: Item Name
          type: string
      required:
      - item_id
      - item_name
      title: GetItem
      type: object
    GetItems:
      properties:
        items:
          items:
            $ref: '#/components/schemas/GetItem'
          title: Items
          type: array
      required:
      - items
      title: GetItems
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    UpdateItem-Input:
      properties:
        item_name:
          title: Item Name
          type: string
      required:
      - item_name
      title: UpdateItem
      type: object
    UpdateItem-Output:
      properties:
        item_id:
          title: Item Id
          type: integer
        item_name:
          title: Item Name
          type: string
      required:
      - item_id
      - item_name
      title: UpdateItem
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Template for FastAPI projects
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
  /items:
    get:
      operationId: get_items_items_get
      parameters:
      - description: The IDs of the items to get
        example: item_ids=1&item_ids=2&item_ids=3
        in: query
        name: item_ids
        required: true
        schema:
          description: The IDs of the items to get
          items:
            type: integer
          required: false
          title: Item IDs
          type: array
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItems'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Items
      tags:
      - Items
    post:
      operationId: create_item_items_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItem-Input'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateItem-Output'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Item
      tags:
      - Items
  /items/{item_id}:
    delete:
      operationId: delete_item_items__item_id__delete
      parameters:
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Item
      tags:
      - Items
    get:
      operationId: get_item_items__item_id__get
      parameters:
      - description: The ID of the item to get
        example: 1
        in: path
        name: item_id
        required: true
        schema:
          description: The ID of the item to get
          exclusiveMinimum: 0
          title: Item ID
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItem'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Item
      tags:
      - Items
    put:
      operationId: update_item_items__item_id__put
      parameters:
      - description: The ID of the item to get
        example: 1
        in: path
        name: item_id
        required: true
        schema:
          description: The ID of the item to get
          exclusiveMinimum: 0
          title: Item ID
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItem-Input'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateItem-Output'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Item
      tags:
      - Items